requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
asttokens=3.0.0=pypi_0
blas=1.0=mkl
brotlicffi=1.0.9.2=py310h5da7b33_1
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.9.9=haa95532_0
certifi=2025.10.5=py310haa95532_0
cffi=2.0.0=py310h02ab6af_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
cloudpickle=3.1.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.3=pypi_0
contourpy=1.3.2=pypi_0
cuda-cccl=13.0.85=h5221881_1
cuda-cccl_win-64=13.0.85=hc667259_1
cuda-cudart=12.1.105=0
cuda-cudart-dev=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-libraries-dev=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvrtc-dev=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=13.0.85=h17533db_0
cuda-opencl-dev=13.0.85=hde976d8_0
cuda-profiler-api=13.0.85=h5221881_0
cuda-runtime=12.1.0=0
cuda-version=13.0=3
cycler=0.12.1=pypi_0
debugpy=1.8.17=pypi_0
decorator=5.2.1=pypi_0
exceptiongroup=1.3.0=pypi_0
executing=2.2.1=pypi_0
expat=2.7.1=h8ddb27b_0
filelock=3.17.0=py310haa95532_0
fonttools=4.60.1=pypi_0
freetype=2.13.3=h0620614_0
giflib=5.2.2=h7edc060_0
gmp=6.3.0=h537511b_0
gmpy2=2.2.1=py310h827c3e9_0
idna=3.7=py310haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.31.0=pypi_0
ipython=8.37.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=py310haa95532_0
joblib=1.5.2=pypi_0
jpeg=9f=ha349fce_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.9.1=pypi_0
khronos-opencl-icd-loader=2024.05.08=h8cc25b3_0
kiwisolver=1.4.9=pypi_0
lcms2=2.16=hb4a4139_0
lerc=3.0=hd77b12b_0
libcublas=12.1.0.26=0
libcublas-dev=12.1.0.26=0
libcufft=11.0.2.4=0
libcufft-dev=11.0.2.4=0
libcurand=10.4.0.35=h86a452a_1
libcurand-dev=10.4.0.35=hde976d8_1
libcusolver=11.4.4.55=0
libcusolver-dev=11.4.4.55=0
libcusparse=12.0.2.55=0
libcusparse-dev=12.0.2.55=0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_1
libjpeg-turbo=2.0.0=h196d8e1_0
libnpp=12.0.2.50=0
libnpp-dev=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjitlink-dev=12.1.105=0
libnvjpeg=12.1.1.14=0
libnvjpeg-dev=12.1.1.14=0
libpng=1.6.50=h46444df_0
libtiff=4.5.1=hd77b12b_0
libuv=1.48.0=h827c3e9_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h3d04722_1
libzlib=1.3.1=h02ab6af_0
lz4-c=1.9.4=h2bbff1b_1
markupsafe=3.0.2=py310h827c3e9_0
matplotlib=3.10.7=pypi_0
matplotlib-inline=0.2.1=pypi_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py310h827c3e9_2
mkl_fft=1.3.11=py310h827c3e9_0
mkl_random=1.2.8=py310hc64d2fc_0
mpc=1.3.1=h827c3e9_0
mpfr=4.2.1=h56c3642_0
mpmath=1.3.0=py310haa95532_0
nest-asyncio=1.6.0=pypi_0
networkx=3.4.2=py310haa95532_0
numpy=2.0.1=py310h055cbcc_1
numpy-base=2.0.1=py310h65a83cf_1
openjpeg=2.5.2=hae555c5_0
openssl=3.0.18=h543e019_0
packaging=25.0=pypi_0
parso=0.8.5=pypi_0
pillow=11.1.0=py310h096bfcc_0
pip=25.2=pyhc872135_1
platformdirs=4.5.0=pypi_0
prompt-toolkit=3.0.52=pypi_0
psutil=7.1.2=pypi_0
pure-eval=0.2.3=pypi_0
pycparser=2.23=py310haa95532_0
pygments=2.19.2=pypi_0
pyparsing=3.2.5=pypi_0
pysocks=1.7.1=py310haa95532_0
python=3.10.19=h981015d_0
python-dateutil=2.9.0.post0=pypi_0
pytorch=2.5.1=py3.10_cuda12.1_cudnn9_0
pytorch-cuda=12.1=hde6ce7c_6
pytorch-mutex=1.0=cuda
pyyaml=6.0.2=py310h827c3e9_0
pyzmq=27.1.0=pypi_0
requests=2.32.5=py310haa95532_0
scikit-learn=1.7.2=pypi_0
scipy=1.15.3=pypi_0
setuptools=80.9.0=py310haa95532_0
six=1.17.0=pypi_0
spyder-kernels=2.4.4=pypi_0
sqlite=3.50.2=hda9a48d_1
stack-data=0.6.3=pypi_0
sympy=1.14.0=py310haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=hf199647_0
torchaudio=2.5.1=pypi_0
torchvision=0.20.1=pypi_0
tornado=6.5.2=pypi_0
traitlets=5.14.3=pypi_0
typing_extensions=4.15.0=py310haa95532_0
tzdata=2025b=h04d1e81_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.5.0=py310haa95532_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
wcwidth=0.2.14=pypi_0
wheel=0.45.1=py310haa95532_0
win_inet_pton=1.1.0=py310haa95532_0
xz=5.6.4=h4754444_1
yaml=0.2.5=he774522_0
zlib=1.3.1=h02ab6af_0
zstd=1.5.7=h56299aa_0
